functions/requirements.txt
firebase_functions~=0.1.0
google-auth
requests
pydantic>=2.0.0
google-cloud-secret-manager